requirements.txt
appdirs==1.4.3
numpy==1.12.1
packaging==16.8
pyparsing==2.2.0
PyQt5==5.8.2
pyqtgraph==0.10.0
pyserial==3.3
sip==4.19.2
six==1.10.0
